Add free places

+++ src/main/kotlin/ru/depo/api/contract/ContractController.kt
@@ -22,7 +22,7 @@ class ContractController(
     @GetMapping("add")
     fun addContract(
         @RequestParam contractdate: String,
-        @RequestParam cost: Float,
+        @RequestParam cost: Long,
         @RequestParam number: String,
         @RequestParam clientid: Long,
         @RequestParam managerid: Long
